sMove all caching logic into a Cache struct
