+++ src/compute/mod.rs
@@ -7,10 +7,9 @@ pub(crate) mod leaf;
 #[cfg(feature = "grid")]
 pub(crate) mod grid;
 
-use crate::data::CACHE_SIZE;
 use crate::error::TaffyError;
 use crate::geometry::{Point, Size};
-use crate::layout::{Cache, Layout, RunMode, SizeAndBaselines, SizingMode};
+use crate::layout::{Layout, RunMode, SizeAndBaselines, SizingMode};
 use crate::node::{Node, Taffy};
 use crate::style::{AvailableSpace, Display};
 use crate::tree::LayoutTree;
@@ -112,9 +111,8 @@ fn compute_node_layout(
 
     // First we check if we have a cached result for the given input
     let cache_run_mode = if child_count == 0 { RunMode::PeformLayout } else { run_mode };
-    if let Some(cached_size_and_baselines) =
-        compute_from_cache(&tree.nodes[node].size_cache, known_dimensions, available_space, cache_run_mode, sizing_mode)
-    {
+    let cached_size = tree.nodes[node].cache.get(known_dimensions, available_space, cache_run_mode, sizing_mode);
+    if let Some(cached_size_and_baselines) = cached_size {
         #[cfg(feature = "debug")]
         NODE_LOGGER.labelled_debug_log("CACHE", cached_size_and_baselines.size);
         #[cfg(feature = "debug")]
@@ -196,13 +194,7 @@ fn compute_node_layout(
     };
 
     // Cache result
-    let cache_slot = compute_cache_slot(known_dimensions, available_space);
-    tree.nodes[node].size_cache[cache_slot] = Some(Cache {
-        known_dimensions,
-        available_space,
-        run_mode: cache_run_mode,
-        cached_size_and_baselines: computed_size_and_baselines,
-    });
+    tree.nodes[node].cache.store(known_dimensions, available_space, cache_run_mode, computed_size_and_baselines);
 
     #[cfg(feature = "debug")]
     NODE_LOGGER.labelled_debug_log("RESULT", computed_size_and_baselines.size);
@@ -225,93 +217,6 @@ fn debug_log_node(
     NODE_LOGGER.labelled_debug_log("known_dimensions", known_dimensions);
     NODE_LOGGER.labelled_debug_log("parent_size", parent_size);
     NODE_LOGGER.labelled_debug_log("available_space", available_space);
-}
-
-/// Return the cache slot to cache the current computed result in
-///
-/// ## Caching Strategy
-///
-/// We need multiple cache slots, because a node's size is often queried by it's parent multiple times in the course of the layout
-/// process, and we don't want later results to clobber earlier ones.
-///
-/// The two variables that we care about when determining cache slot are:
-///
-///   - How many "known_dimensions" are set. In the worst case, a node may be called first with neither dimensions known known_dimensions,
-///     then with one dimension known (either width of height - which doesn't matter for our purposes here), and then with both dimensions
-///     known.
-///   - Whether unknown dimensions are being sized under a min-content or a max-content available space constraint (definite available space
-///     shares a cache slot with max-content because a node will generally be sized under one or the other but not both).
-///
-/// ## Cache slots:
-///
-/// - Slot 0: Both known_dimensions were set
-/// - Slot 1: 1 of 2 known_dimensions were set and the other dimension was either a MaxContent or Definite available space constraint
-/// - Slot 2: 1 of 2 known_dimensions were set and the other dimension was a MinContent constraint
-/// - Slot 3: Neither known_dimensions were set and we are sizing under a MaxContent or Definite available space constraint
-/// - Slot 4: Neither known_dimensions were set and we are sizing under a MinContent constraint
-#[inline]
-fn compute_cache_slot(known_dimensions: Size<Option<f32>>, available_space: Size<AvailableSpace>) -> usize {
-    let has_known_width = known_dimensions.width.is_some();
-    let has_known_height = known_dimensions.height.is_some();
-
-    // Slot 0: Both known_dimensions were set
-    if has_known_width && has_known_height {
-        return 0;
-    }
-
-    // Slot 1: width but not height known_dimension was set and the other dimension was either a MaxContent or Definite available space constraint
-    // Slot 2: width but not height known_dimension was set and the other dimension was a MinContent constraint
-    if has_known_width && !has_known_height {
-        return 1 + (available_space.height == AvailableSpace::MinContent) as usize;
-    }
-
-    // Slot 3: height but not width known_dimension was set and the other dimension was either a MaxContent or Definite available space constraint
-    // Slot 4: height but not width known_dimension was set and the other dimension was a MinContent constraint
-    if !has_known_width && has_known_height {
-        return 3 + (available_space.width == AvailableSpace::MinContent) as usize;
-    }
-
-    // Slot 5: Neither known_dimensions were set and we are sizing under a MaxContent or Definite available space constraint
-    // Slot 6: Neither known_dimensions were set and we are sizing under a MinContent constraint
-    5 + (available_space.width == AvailableSpace::MinContent) as usize
-}
-
-/// Try to get the computation result from the cache.
-#[inline]
-fn compute_from_cache(
-    cache: &[Option<Cache>; CACHE_SIZE],
-    known_dimensions: Size<Option<f32>>,
-    available_space: Size<AvailableSpace>,
-    run_mode: RunMode,
-    sizing_mode: SizingMode,
-) -> Option<SizeAndBaselines> {
-    for entry in cache.iter().flatten() {
-        // Cached ComputeSize results are not valid if we are running in PerformLayout mode
-        if entry.run_mode == RunMode::ComputeSize && run_mode == RunMode::PeformLayout {
-            return None;
-        }
-
-        let cached_size = entry.cached_size_and_baselines.size;
-
-        if (known_dimensions.width == entry.known_dimensions.width || known_dimensions.width == Some(cached_size.width))
-            && (known_dimensions.height == entry.known_dimensions.height
-                || known_dimensions.height == Some(cached_size.height))
-            && (known_dimensions.width.is_some()
-                || entry.available_space.width.is_roughly_equal(available_space.width)
-                || (sizing_mode == SizingMode::ContentSize
-                    && available_space.width.is_definite()
-                    && available_space.width.unwrap() >= cached_size.width))
-            && (known_dimensions.height.is_some()
-                || entry.available_space.height.is_roughly_equal(available_space.height)
-                || (sizing_mode == SizingMode::ContentSize
-                    && available_space.height.is_definite()
-                    && available_space.height.unwrap() >= cached_size.height))
-        {
-            return Some(entry.cached_size_and_baselines);
-        }
-    }
-
-    None
 }
 
 /// Creates a layout for this node and its children, recursively.
