+++ src/data.rs
@@ -2,11 +2,9 @@
 //!
 //! Used to compute layout for Taffy trees
 //!
-use crate::layout::{Cache, Layout};
+use crate::cache::Cache;
+use crate::layout::Layout;
 use crate::style::Style;
-
-/// The number of cache entries for each node in the tree
-pub(crate) const CACHE_SIZE: usize = 7;
 
 /// Layout information for a given [`Node`](crate::node::Node)
 ///
@@ -20,15 +18,15 @@ pub(crate) struct NodeData {
     /// Should we try and measure this node?
     pub(crate) needs_measure: bool,
 
-    /// The primary cached results of the layout computation
-    pub(crate) size_cache: [Option<Cache>; CACHE_SIZE],
+    /// The cached results of the layout computation
+    pub(crate) cache: Cache,
 }
 
 impl NodeData {
     /// Create the data for a new node
     #[must_use]
     pub const fn new(style: Style) -> Self {
-        Self { style, size_cache: [None; CACHE_SIZE], layout: Layout::new(), needs_measure: false }
+        Self { style, cache: Cache::new(), layout: Layout::new(), needs_measure: false }
     }
 
     /// Marks a node and all of its parents (recursively) as dirty
@@ -36,6 +34,6 @@ impl NodeData {
     /// This clears any cached data and signals that the data must be recomputed.
     #[inline]
     pub fn mark_dirty(&mut self) {
-        self.size_cache = [None; CACHE_SIZE];
+        self.cache.clear();
     }
 }
